create group with default layout and more

+++ src/messages/messages.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Param, Put, UseGuards } from '@nestjs/common'
+import { Body, Controller, Delete, ForbiddenException, NotFoundException, Param, Put, UseGuards } from '@nestjs/common'
 import { CurrentUser } from 'src/auth/decorators/current-user.decorator'
 import { JwtAuthGuard } from 'src/auth/guards/jwt.guard'
 import { RequestResponseUser } from 'src/auth/types/request-response'
@@ -6,7 +6,6 @@ import { ChatMessageService } from './chat-message.service'
 import { MessagesService } from './messages.service'
 import { ChatMessage } from './models/chat-message'
 import { Message } from './models/messages.model'
-import { messageAuthorValidate } from './utils/message-author-validate'
 
 
 @Controller('/messages')
@@ -15,7 +14,7 @@ export class MessagesController {
     constructor(
         private messagesService: MessagesService,
         private chatMessageService: ChatMessageService,
-        // private channelMessagesService: ChannelMessagesService,
+        // private channelMessagesService: TextChannelMessagesService,
     ) {}
 
     @Put('/:id')
@@ -26,7 +25,10 @@ export class MessagesController {
         @Body() { text }: { text: string }
     ): Promise<Message> {
         const message: Message = await this.messagesService.getMessageById(messageId)
-        messageAuthorValidate(message.userId, user.id)
+        if (!message)
+            throw new NotFoundException({ message: 'Message not found' })
+        if (message.userId !== user.id)
+            throw new ForbiddenException({ message: 'You have no access' })
         const updatedMessage: Message = await this.messagesService.updateMessage({
             message,
             text
@@ -41,11 +43,14 @@ export class MessagesController {
         @Param('id') messageId: number
     ): Promise<Message> {
         const message: Message = await this.messagesService.getMessageById(messageId)
-        messageAuthorValidate(message.userId, user.id)
+        if (!message)
+            throw new NotFoundException({ message: 'Message not found' })
+        if (message.userId !== user.id)
+            throw new ForbiddenException({ message: 'You have no access' })
         await this.messagesService.deleteMessage({ message })
         const messageInChat: ChatMessage = await this.chatMessageService.getChatMessageColumn(message.id)
         // if (!messageInChat)
-        //     await this.channelMessagesService.getChannelMessageColumn(message.id)
+        //     await this.channelMessagesService.getTextChannelMessageColumn(message.id)
         return message
     }
